Fix -Wextra warnings
git-svn-id: svn+ssh://svn.code.sf.net/p/mingw-w64/code/trunk@1198 4407c894-4637-0410-b4f5-ada5f102cad1

+++ mingw-w64-crt/gdtoa/dmisc.c
@@ -41,7 +41,7 @@ char *rv_alloc (int i)
 
 	j = sizeof(ULong);
 	for(k = 0;
-		sizeof(Bigint) - sizeof(ULong) - sizeof(int) + j <= i;
+		(int) (sizeof(Bigint) - sizeof(ULong) - sizeof(int)) + j <= i;
 		j <<= 1)
 			k++;
 	r = (int*)Balloc(k);
